requirements.txt
# Django and API framework
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0

# Machine Learning and NLP
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Data processing
pandas>=2.0.0

# Database
psycopg2-binary>=2.9.0

# Testing
pytest>=7.4.0
pytest-django>=4.5.0
pytest-cov>=4.1.0

# Production
gunicorn>=21.2.0
whitenoise>=6.5.0

# Development
python-decouple>=3.8 